+++ src/main/java/com/forgestove/create_cyber_goggles/CCGConfig.java
@@ -49,6 +49,7 @@ public class CCGConfig implements ConfigData {
 		@Tooltip public boolean removeCooldown = true;
 	}
 	public static class Other {
+		@Tooltip public boolean rightClickPenetrate = false;
 		@Tooltip public boolean rainbowDebug = false;
 		@Tooltip public boolean forcedBackend = false;
 		@Tooltip @RequiresRestart public boolean nonrandomScrap = true;